+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.4 (June 27, 2018)
 - Add SQL counter to `fly\Database`.
+- Add query cache to `fly\Database`.
 - Add function support in select part to Query Builder.
 - Add 'NOT IN', 'LIKE', 'NOT LIKE' support to Query Builder.
 - Add multi file support for config in class `fly\Route`.